requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0

# Testing and development
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.5b2
flake8>=3.9.2
mypy>=0.910

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
nbformat>=5.1.3

# Optional dependencies
tqdm>=4.61.0  # For progress bars
PyYAML>=5.4.1  # For configuration files
h5py>=3.2.1   # For HDF5 file support
openpyxl>=3.0.7  # For Excel file support
requests>=2.25.1  # For API integrations

# Documentation
sphinx>=4.0.2
sphinx-rtd-theme>=0.5.2
nbsphinx>=0.8.6

# Type checking
types-PyYAML>=5.4.3
types-requests>=2.25.0

# CLI Tools
click>=8.0.0  # For command-line interface